requirements.txt
playwright
requests
streamlit
pypdf
groq
supabase
streamlit-calendar
python-dotenv